Don't indirectly link dependent libraries.
The default behaviour for ld allows users to 'indirectly' link to required
objects/libraries through intermediate objects/libraries. While this is
convenient, it can also be dangerous because it makes your program's
dependencies tied to the dependencies of other objects.

Beginning with Fedora 13 this will be changed and you need to explicitly
link all dependent libraries.

More details can be found here:
http://fedoraproject.org/wiki/UnderstandingDSOLinkChange

This patch fixes all such cases in strongSwan.