libcharon/sa/ikev1/tasks/isakmp_delete: Reopen deleted IKE_SAs.
If an IKE peer sends a delete message for the ISAKMP association, we
would previously just delete the entire association and remain that way
until the peer either re-established a new connection to us or the user
administratively brought it back up. We now check if the IKE we're
deleting is in an ESTABLISHED state prior to deletion and initiate a new
one if so.